Migrate build.scala to build.sbt
- add .idea/compiler.xml (since I didn't add linting/fixes for tests, so
  disabled there)
- add sbt-native-packager
- remove sbt-scalastyle, replacing it with scapegoat and lintr
  (scapegoat has to be run manually)
- fix all lint errors (not scapegoat yet) in compiled files (not tests)

+++ src/main/scala/mesosphere/marathon/api/ResponseFacade.scala
@@ -15,7 +15,7 @@ class ResponseFacade extends HttpResponse {
   }
   override def cookie(name: String, value: String, maxAge: Int, secure: Boolean): Unit = {
     //scalastyle:off null
-    builder.cookie(new NewCookie(name, value, null, null, null, maxAge.toInt, secure))
+    builder.cookie(new NewCookie(name, value, null, null, null, maxAge, secure))
   }
   override def body(mediaType: String, bytes: Array[Byte]): Unit = {
     builder.`type`(mediaType)
